fix escaped packet size limitation

+++ src/SerialProtocol/SerialProtocol.h
@@ -69,7 +69,7 @@ private:
     SerialParser parser;
     uint8_t packet[MAX_PACKET_SIZE];
     uint8_t escaped_packet[MAX_PACKET_SIZE*2];
-    uint8_t escape_packet(uint8_t* data, size_t len);
+    size_t escape_packet(uint8_t* data, size_t len);
     uint8_t crc8(const uint8_t* data, size_t len);
     void parse_serial();
 };
